# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IResponse_any_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{},
							{
								"items": {},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralFileResponseDto": {
				"properties": {
					"fileGuid": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fileGuid",
					"fileName",
					"fileUrl",
					"contentType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralFileDto": {
				"properties": {
					"fileName": {
						"type": "string"
					},
					"fileData": {
						"type": "string"
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fileName",
					"fileData",
					"contentType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetNotificationDto": {
				"properties": {
					"notificationId": {
						"type": "number",
						"format": "double"
					},
					"notificationGuid": {
						"type": "string"
					},
					"userGuid": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"data": {},
					"isRead": {
						"type": "boolean"
					},
					"createdBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"notificationId",
					"notificationGuid",
					"userGuid",
					"title",
					"body",
					"data",
					"isRead",
					"createdBy",
					"createdDateTime",
					"updatedBy",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetNotificationDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetNotificationDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetNotificationDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetNotificationDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetNotificationDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateNoticeDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/GeneralFileDto"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"startDate",
					"endDate",
					"attachments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetNoticeDto": {
				"properties": {
					"noticeId": {
						"type": "number",
						"format": "double"
					},
					"noticeGuid": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"noticeId",
					"noticeGuid",
					"title",
					"description",
					"startDate",
					"endDate",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetNoticeDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetNoticeDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetNoticeDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetNoticeDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetNoticeDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetNoticeDetailsDto": {
				"properties": {
					"noticeId": {
						"type": "number",
						"format": "double"
					},
					"noticeGuid": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/GeneralFileResponseDto"
						},
						"type": "array"
					},
					"status": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"noticeId",
					"noticeGuid",
					"title",
					"description",
					"startDate",
					"endDate",
					"attachments",
					"status",
					"createdBy",
					"createdDateTime",
					"updatedBy",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetNoticeDetailsDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetNoticeDetailsDto"
							},
							{
								"items": {
									"$ref": "#/components/schemas/GetNoticeDetailsDto"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EditNoticeDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"deletedAttachments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"newAttachments": {
						"items": {
							"$ref": "#/components/schemas/GeneralFileDto"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetFacilityBookingHistoryDto": {
				"properties": {
					"bookingId": {
						"type": "number",
						"format": "double"
					},
					"bookingGuid": {
						"type": "string"
					},
					"facilityName": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"bookedBy": {
						"type": "string"
					},
					"isCancelled": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"bookingId",
					"bookingGuid",
					"facilityName",
					"startDate",
					"endDate",
					"bookedBy",
					"isCancelled",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetFacilityBookingHistoryDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetFacilityBookingHistoryDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetFacilityBookingHistoryDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetFacilityBookingHistoryDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetFacilityBookingHistoryDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Secure Guard Pro",
		"version": "1.0.0",
		"description": "Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/notifications": {
			"get": {
				"operationId": "getNotification",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetNotificationDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNotificationDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notification"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"STF"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notices/admin": {
			"post": {
				"operationId": "createNotice",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNoticeDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "getNoticeByAdmin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetNoticeDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "direction",
						"required": true,
						"schema": {}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notices/admin/{id}/details": {
			"get": {
				"operationId": "getNoticeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notices/admin/{id}": {
			"put": {
				"operationId": "editNoticeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditNoticeDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteNoticeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice Management"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA",
							"STF"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notices": {
			"get": {
				"operationId": "getNotice",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetNoticeDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notices/{id}/details": {
			"get": {
				"operationId": "getNoticeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetNoticeDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA",
							"STF"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facilities/admin/bookings": {
			"get": {
				"operationId": "getFacilityBookingHistoryByAdmin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetFacilityBookingHistoryDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetFacilityBookingHistoryDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Facility Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "direction",
						"required": true,
						"schema": {}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}